final_project/requirements.txt
Flask==2.1.2
Flask-WTF==1.0.1
ibm-watson==6.0.0
python-dotenv==0.20.0
pyJWT==2.4.0
